Add some text connecting intro to second section
I've created a first-level header for the second of the section (currently "Data for studying the human impacts of tropical cyclones", but that's a bit of a placeholder---we can come up with a better section title later.

I'm also doing some edits to try to ease a bit more from the Intro to the second section, including adding some text to introduce the second section. Some of this might end up being too repetative with the end of the Intro, and so we likely will trim or maybe completely edit out when we get to later edits.